fix: refactoring gf-344
