+++ apps/frontend/src/pages/contributors/libs/components/contributor-merge-form/libs/helpers/helpers.ts
@@ -1,1 +1,1 @@
-export { getContributorOptions } from "./get-contributor-options.helper.js";
+export { getContributorOptions } from "./get-contributor-options/get-contributor-options.helper.js";
