+++ packages/shared/src/modules/contributors/libs/validation-schemas/validation-schemas.ts
@@ -1,2 +1,2 @@
 export { contributorMerge } from "./contributor-merge.validation-schema.js";
-export { contributorPatch } from "./contributor-patch.validation-schemas.js";
+export { contributorPatch } from "./contributor-patch.validation-schema.js";
